requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.0.1
opencv-python==4.11.0.86
numpy==2.2.3
moviepy==1.0.3
openai==1.65.4
python-dotenv==1.0.1
gTTS==2.5.0
requests==2.32.3
pydantic==2.6.3 